modules/usrEnv: allow desktop options to specify their own packages
This should make it easier to share desktop packages inbetween desktops. This is a problem since I get Hyprland package from the flake inputs, and don't want to type the package name every single time. Currently, the Hyprland package is duplicated since I depend on it for my gamemoderun script.